Build windowjs-p5 with the merged p5 loader and library
