+++ src/tools/CMakeLists.txt
@@ -9,3 +9,7 @@ add_executable(embed
 
 target_include_directories(embed PRIVATE ../../libraries/v8/third_party/zlib)
 target_link_libraries(embed PRIVATE v8)
+
+add_executable(merge-p5
+    merge_p5.cc
+)
